Add priority addition for certain game sprites
This allows us to make the player sprite display above other sprites.
Previously the Z-index of our sprites was tied to the order they were
added. This addresses that by allowing us to explicitly pass an
additional parameter telling the game world to put that sprite in
a high priority stage.