WIP: Add doors
This modifies classes to introduce the concept of entities with
destination, as well as some logic to create a door in the first
area.

This does not yet include logic to only display things in the current
area, to prevent enemies from walking over doors, or to move the player
to a new area when they walk over a door.