Prevent player from getting deleted when walking on walls
This bug happened because we moved from query.filter, where we can
early return with nothing, since the function returns void, to
query.map where we return an object diffing keys of the model. We were
returning undefined, which meant removing the entity from the ecs system
altogether, which wasn't the behavior we wanted.